+++ src/main/java/com/karifovas/gamerating/dto/FactorDto.java
@@ -2,10 +2,12 @@ package com.karifovas.gamerating.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record FactorDto(
         String id,
         String name,
         String description,
-        Float value
+        BigDecimal value
 ) { }
