+++ src/main/java/com/karifovas/gamerating/dto/RatingDto.java
@@ -1,8 +1,8 @@
 package com.karifovas.gamerating.dto;
 
-import com.karifovas.gamerating.model.Rating;
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Builder
@@ -13,5 +13,5 @@ public record RatingDto(
         String description,
         List<FactorDto> factors,
         List<AdjustmentDto> adjustments,
-        Float value
+        BigDecimal value
 ) { }
